+++ packages/vue/src/components/factory.ts
@@ -51,6 +51,10 @@ export type HTMLPolymorphicProps<T extends ElementType> = Omit<
 
 export type HTMLArkProps<T extends DOMElements> = HTMLPolymorphicProps<T>
 
+// For self closing tags, don't provide default slots because of hydration issue
+const SELF_CLOSING_TAGS = 'br, hr, img, input, area, textarea'.split(', ')
+const isSelfClosingTag = (tag: unknown) => typeof tag === 'string' && SELF_CLOSING_TAGS.includes(tag)
+
 const withAsChild = (component: ElementType) => {
   return defineComponent({
     name: 'Polymorphic',
@@ -62,29 +66,21 @@ const withAsChild = (component: ElementType) => {
       },
     },
     setup(props, { attrs, slots }) {
-      if (!props.asChild) return () => h(component, { ...attrs }, slots.default?.())
-      return () => h(Dynamic, attrs, { default: slots.default })
+      if (!props.asChild) return () => h(component, attrs, isSelfClosingTag(component) ? undefined : slots.default?.())
+      return () => h(Dynamic, attrs, slots)
     },
   })
 }
 
 export function jsxFactory() {
-  const cache = new Map()
-
-  const factory = new Proxy(withAsChild, {
+  return new Proxy(withAsChild, {
     apply(_target, _thisArg, argArray) {
       return withAsChild(argArray[0])
     },
     get(_, element) {
-      const asElement = element as unknown as ElementType
-      if (!cache.has(asElement)) {
-        cache.set(asElement, withAsChild(asElement))
-      }
-      return cache.get(asElement)
+      return withAsChild(element as ElementType)
     },
   }) as unknown as HTMLPolymorphicComponents
-
-  return factory
 }
 
 export const ark = jsxFactory()