+++ templates/nuxt/app/app.vue
@@ -1,8 +1,10 @@
 <script setup lang="ts">
 import Avatar from './components/avatar.vue'
+import FieldDemo from './components/field-demo.vue'
 </script>
 
 <template>
   <h1>Welcome to Ark UI!</h1>
   <Avatar name="Christian Schröter" src="https://avatars3.githubusercontent.com/u/1846056" />
+  <FieldDemo />
 </template>