+++ packages/vue/CHANGELOG.md
@@ -5,6 +5,10 @@
 - **Fieldset**: Reverted `Fieldset.Legend` to render as a `legend` element instead of a `div` element (reverting changes
   from v5.21.0) for better semantic HTML
 
+### Fixed
+
+- **Field**: Fixed hydration issues when using `Field.Input` or `Field.Textarea` components
+
 ## [5.23.0] - 2025-09-08
 
 ### Added